+++ lib/src/runner.dart
@@ -8,9 +8,9 @@ import 'package:args/command_runner.dart';
 import 'package:path/path.dart' as p;
 import 'package:term_glyph/term_glyph.dart' as glyph;
 
-import 'src/io.dart';
-import 'src/migrators/division.dart';
-import 'src/migrators/module.dart';
+import 'io.dart';
+import 'migrators/division.dart';
+import 'migrators/module.dart';
 
 /// A command runner that runs a migrator based on provided arguments.
 class MigratorRunner extends CommandRunner<Map<Uri, String>> {