+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.0.0-alpha.5
 
+* **Breaking change**: Remove `lib/runner.dart`. This package is only meant to
+  be used as an executable, not as a library.
+
 ### Division Migrator
 
 * Fix a type error.